+++ DutWrapper.Tester/NewsTest.cs
@@ -22,8 +22,6 @@ namespace DutWrapper.Tester
                 var data = NewsInstance.GetNewsGlobal(i).Result;
                 if (data == null)
                     throw new NullReferenceException($"Internal error from function. Did you connected the internet?");
-                if (data.Count == 0)
-                    throw new NullReferenceException($"No datas in page {i}. Did you connected the internet?");
 
                 news.AddRange(data);
             }
@@ -47,8 +45,6 @@ namespace DutWrapper.Tester
                 var data = NewsInstance.GetNewsSubject(i).Result;
                 if (data == null)
                     throw new NullReferenceException($"Internal error from function. Did you connected the internet?");
-                if (data.Count == 0)
-                    throw new NullReferenceException($"No datas in page {i}. Did you connected the internet?");
 
                 news.AddRange(data);
             }
@@ -60,5 +56,76 @@ namespace DutWrapper.Tester
                 Encoder = System.Text.Encodings.Web.JavaScriptEncoder.UnsafeRelaxedJsonEscaping
             }));
         }
+
+        [TestMethod]
+        public void GetNews_StudentAffairs()
+        {
+            int NEWS_COUNT = 5;
+
+            List<NewsGlobal> news = new List<NewsGlobal>();
+            for (int i = 1; i <= NEWS_COUNT; i++)
+            {
+                var data = NewsInstance.GetNewsStudentAffairs(i).Result;
+                if (data == null)
+                    throw new NullReferenceException($"Internal error from function. Did you connected the internet?");
+
+                news.AddRange(data);
+            }
+
+            Debug.WriteLine($"Total news in {NEWS_COUNT} page(s): {news.Count}");
+            Debug.WriteLine(JsonSerializer.Serialize<List<NewsGlobal>>(news, new JsonSerializerOptions
+            {
+                WriteIndented = true,
+                Encoder = System.Text.Encodings.Web.JavaScriptEncoder.UnsafeRelaxedJsonEscaping
+            }));
+        }
+
+        [TestMethod]
+        public void GetNews_Examination()
+        {
+            int NEWS_COUNT = 5;
+
+            List<NewsGlobal> news = new List<NewsGlobal>();
+            for (int i = 1; i <= NEWS_COUNT; i++)
+            {
+                var data = NewsInstance.GetNewsExamination(i).Result;
+                if (data == null)
+                    throw new NullReferenceException($"Internal error from function. Did you connected the internet?");
+
+                news.AddRange(data);
+            }
+
+            Debug.WriteLine($"Total news in {NEWS_COUNT} page(s): {news.Count}");
+            Debug.WriteLine(JsonSerializer.Serialize<List<NewsGlobal>>(news, new JsonSerializerOptions
+            {
+                WriteIndented = true,
+                Encoder = System.Text.Encodings.Web.JavaScriptEncoder.UnsafeRelaxedJsonEscaping
+            }));
+        }
+
+        [TestMethod]
+        public void GetNews_TuitionFee()
+        {
+            int NEWS_COUNT = 5;
+
+            List<NewsGlobal> news = new List<NewsGlobal>();
+            for (int i = 1; i <= NEWS_COUNT; i++)
+            {
+                var data = NewsInstance.GetNewsTuitionFee(i).Result;
+                if (data == null)
+                    throw new NullReferenceException($"Internal error from function. Did you connected the internet?");
+
+                news.AddRange(data);
+            }
+
+            Debug.WriteLine($"Total news in {NEWS_COUNT} page(s): {news.Count}");
+            Debug.WriteLine(JsonSerializer.Serialize<List<NewsGlobal>>(news, new JsonSerializerOptions
+            {
+                WriteIndented = true,
+                Encoder = System.Text.Encodings.Web.JavaScriptEncoder.UnsafeRelaxedJsonEscaping
+            }));
+        }
+
+
     }
 }